refactor: smaller changes/tests
