+++ pkg/providers/azclient/aksmachinesheaderbatch/client.go
@@ -18,6 +18,7 @@ package aksmachinesheaderbatch
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice/v9"
@@ -38,6 +39,18 @@ type AKSMachinesHeaderBatchAPI interface {
 	// - The fact that HandlableError is considered one of the "expected" states in this
 	//   context, just not ideal. Operational error, on the other hand, is more of a bug.
 	BeginCreateWithBatch(ctx context.Context, resourceGroupName string, resourceName string, agentPoolName string, aksMachineName string, machine *armcontainerservice.Machine) (*offerings.HandlableError, error)
+}
+
+// HandlableError is a code + message pair extracted from an API error response.
+// It is intentionally minimal and API-agnostic — the caller decides how to interpret it.
+type HandlableError struct {
+	Code    string
+	Message string
+}
+
+// Implement the error interface so it can be returned as an error type if needed.
+func (e *HandlableError) Error() string {
+	return fmt.Sprintf("%s: %s", e.Code, e.Message)
 }
 
 // We don't need the rest of machine API interface. Just create.
